Put all of console effect example in one file
