+++ samples/Eff.Examples.Console/Program.cs
@@ -1,18 +1,53 @@
 ﻿using System.Threading.Tasks;
+using Nessos.Effects;
+using Nessos.Effects.Handlers;
 
-namespace Nessos.Effects.Examples.Console
+static async Eff Test()
 {
-    partial class Program
+    await ConsoleEffect.Print("Enter your name: ");
+    await ConsoleEffect.Print($"Hello, {await ConsoleEffect.Read()}!\n");
+}
+
+await Test().Run(new ConsoleEffectHandler());
+
+public class ConsolePrintEffect(string message) : Effect
+{
+    public string Message { get; } = message;
+}
+
+public class ConsoleReadEffect : Effect<string?>;
+
+public static class ConsoleEffect
+{
+    public static ConsolePrintEffect Print(string message) => new(message);
+    public static ConsoleReadEffect Read() => new();
+}
+
+/// <summary>
+///   Handles console effects by calling the standard System.Console API
+/// </summary>
+public class ConsoleEffectHandler : EffectHandler
+{
+    public override ValueTask Handle<TResult>(EffectAwaiter<TResult> awaiter)
     {
-        static async Eff Test()
+        switch (awaiter)
         {
-            await ConsoleEffect.Print("Enter your name: ");
-            await ConsoleEffect.Print($"Hello, { await ConsoleEffect.Read()}!\n");
+            case EffectAwaiter { Effect: ConsolePrintEffect effect } awtr:
+            {
+                Console.Write(effect.Message);
+                awtr.SetResult();
+                break;
+            }
+            case EffectAwaiter<string?> { Effect: ConsoleReadEffect } awtr:
+            {
+                string? message = Console.ReadLine();
+                awtr.SetResult(message);
+                break;
+            }
+            default:
+                throw new NotSupportedException(awaiter.Effect.GetType().Name);
         }
 
-        static async Task Main()
-        {
-            await Test().Run(new ConsoleEffectHandler());
-        }
+        return default;
     }
 }
